Specify all job dependencies
The "build" parameter may be not known at the job collecting time, so
use a placeholder build=None dependency that will be later used for job
ordering, but for now ignore such dependencies in check_dependencies. It
can be later improved to resolve build=None into "all build values of a
component", but it's not necessary for now.

With that change, it's possible to specify dependencies for "upload"
jobs (that will depend on all packages), but also "build", "sign" and
"publish" jobs when .qubesbuilder file is not fetched yet.